refactor: fold converter.convert into the conversion module
wrappers/converter.convert had a single production caller (convert_file), so
its seam was hypothetical, not real — and `convert` was overloaded three times
(FileHandler, conversion, converter).

Inline its working-dir + subprocess body as a private _run_tool helper in
tasks/conversion.py, so the whole conversion — run, verify, log-strip — lives in
one module. The real seam that varies (which tool builds the command) stays one
level down at the MediaTool interface in wrappers/tools.py, untouched.

test_converter.py's wiring tests move onto _run_tool inside test_conversion.py
(an internal seam used by the module's own tests); the two convert_file tests
that patched conversion.convert now patch _run_tool. CLAUDE.md updated to point
at the new home.
